server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
gunicorn = "^20.1.0"
djangorestframework-simplejwt = "^5.2.1"
python-decouple = "^3.6"
dj-database-url = "^1.0.0"
psycopg2 = "^2.9.4"
psycopg2-binary = "^2.9.4"
django-cors-headers = "^3.13.0"

[tool.poetry.dev-dependencies]
django-extensions = "^3.2.1"

[tool.poetry.group.dev.dependencies]
drf-yasg = "^1.21.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
